+++ src/main/java/com/swp/autocarwash/booking/port/VehiclePort.java
@@ -31,7 +31,7 @@ public interface VehiclePort {
      * @author Phong
      * @version 1.0
      */
-    List<VehicleContract> getVehiclesByCustomer(Integer customerId);
+    List<VehicleContract> getVehiclesByCustomer(Long customerId);
 
     /**
      *
@@ -47,7 +47,7 @@ public interface VehiclePort {
      * @author Phong
      * @version 1.0
      */
-    Integer getCurrentCustomerId();
+    Long getCurrentCustomerId();
 
     /**
      *
@@ -67,7 +67,7 @@ public interface VehiclePort {
      * @author Phong
      * @version 1.0
      */
-    boolean validateVehicleOwnership(Integer vehicleId, Integer customerId);
+    boolean validateVehicleOwnership(Long vehicleId, Long customerId);
 
     /**
      *
@@ -86,5 +86,5 @@ public interface VehiclePort {
      * @author Phong
      * @version 1.0
      */
-    VehicleContract getById(Integer id);
+    VehicleContract getById(Long id);
 }
